Fix #17: Let teachers change students' teams via the web interface.
Add a new field to the bottom of the edit view, to allow editing of
the team via a dropdown box. This is generated from a generalised
implementation of the code that was originally used for the teams'
dropdown in the registration view, and will show the user's current
team as the pre-selected option.

Since the collection of available teams comes from the editor, not
the user being edited, make their user available to the edit view.
Ensure that the teams collections from the backend are stored in
the user objects for both the logged-in user and the current user
of interest.

Change-Id: I136ea5005150e4fa0c6efd8dd98077d8d9b4e66e