Let users login with any case of username.
This is already the case for our other services, specifically the
IDE and forums, so it's incongruous to prevent it here. We still
respond with the canonical username, and (should) use that internally
so internal references will remain canonical.

The only occasion where we require case matching is the activation
of new accounts, where the username is used as a sqlite lookup key.
In this case, the user is expected to be clicking a link sent in an
email, so the need for tolerance is greatly reduced anyway.

Change-Id: I66e0aa2f60801fcc424c043db55abfb3c9af9fad