Fix UI issue when in multiple colleges
Previously the count of users was a global that was reset for each
college. Since each college then decremented it, it wasn't ever 0
when a college actually finished loading its users (unless only
one college actually had any members, which can't be possible if
the current user is in them all). Make this variable a local so
that each college is handled separately.

While this can result in the status bar showing the wrong text,
that's a small price to pay for the UI actually working properly.

Change-Id: I73a145f2760fdf7da7758d4e8b7fc54b1a568e1c