Fix #3112: Implement media-consent setting backend
This adds support for certain users being able to set the media-
consent status of other users via the /user/:username endpoint.

When changed an email to the modified user is also sent, so that
they know their ticket is available.

Patch set 2: fix test teardown and ensure that media consent is
             only recorded once (thus avoiding duplicate emails)

Change-Id: Ifd4ace4831717cb9babf1e1f76ce5a2d2ad24d5c