Fix #3181: Don't require user-admin to update media-consent
It appears that I missed the check at the top of the user-post
handler that the posting user can administrate the given user.
While we don't want blueshirts to be able to generally update any
user (the relevant team-leader should instead perform the update),
we do need media-consent-admins to be able to set this value for
all users.

Rearrange the checks to allow this while still preventing other
changes as described above. I think we should probably separate
this functionality to a different endpoint (possibly
/user/:username/media-consent) at some point, though that will
need a major-version bump of the API.

Change-Id: Iaeb788d7c5492b93c4925cc70275c767c00f5755