Fix #3181: Extend MCF Blueshirt rights to read user details
This updates libnemesis to allow media consent blueshirts to be able
to read the details of any user so they can confirm they match the
form they have. Also switch to using the pre-created mcf blueshirt
user for the local testing of the write functionality now that
such user exists.

Change-Id: Icef749bd7ae7cdf682a3fad1bd0c828d378c9e5d