Ensure the most current username is always used for highlighting
It is possible to request a refresh of the college lists and then
click on a different user. This can result in the current username
changing after the college fetch has happened. Regardless of
which fetch loads first this can lead to weirdness where the
highlights are out of step with the edit view and location.hash.

It appears that the redundant hash change handling removed in
may have been masking this.

Change-Id: Ie7339ed9672222f624d407db8f7f4393871d1106