Add an overview page for blueshirts
This is primarily aimed at Team Coordinators, though may be useful
for any blueshirt who is a member of more than one team. Anyone
who is a member of more than two teams will now default to this
view, which should be considerably faster to load as it doesn't
try to load all the details for every user.

This also adds links to the new "overview" and existing "list"
pages to the navigation bar.

Change-Id: Iace3f784dc510db4745f568419a386b9a1e903e7